requirements.txt
aiogram
python-dotenv
tortoise-orm
Pillow
aioschedule
aioredis
redis
asyncpg
aiomysql
aerich
requests